requirements.txt
fastapi==0.115.11
langchain==0.3.20
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_ollama==0.2.3
pydantic==2.10.6
python-dotenv==1.0.1
